Fixed bug where Hits.id() returned a Float instead of an int Bug found by Terry Steichen
git-svn-id: https://svn.apache.org/repos/asf/lucene/java/trunk@149892 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/lucene/search/Hits.java
@@ -134,7 +134,7 @@ public final class Hits {
   }
   
  /** Returns the id for the nth document in this set. */ 
-  public final float id(int n) throws IOException {
+  public final int id(int n) throws IOException {
     return hitDoc(n).id;
   }
